feat: ホームタブを (home) Stack 化してヘッダーを追加
NativeTabs 導入 (163bfff) の際に外したヘッダーをホーム画面に復元するため、
`(tabs)/(home)/_layout.tsx` に Stack を新設して設定タブと同じ
"NativeTabs > Stack > 画面" 構造に揃えた。

合わせて、ナビゲーションパターン関連の `index.tsx` / `modal.tsx` /
`form-sheet.tsx` を `(tabs)/(home)/navigation-patterns/` 配下へ移動し、
ホーム → ナビゲーションパターン の遷移を内側 Stack の push に変更。これで
ホーム→ナビゲーションパターンと設定→テーマで「タブバー残ったまま push」
「Stack ヘッダーが push アニメーションで切り替わる」「戻るボタンに親画面の
タイトルがラベル表示される」挙動が完全に一致する。

当初はナビゲーションパターン側を「タブを覆う」現状の挙動のまま残す案も
検討したが、これは「タブ配下 Stack」と「タブ上層 Stack」の挙動比較目的の
構成だった。NativeTabs に移行済みで比較目的を果たしたため、より自然な
タブ内 push に統一する方針を採用。

modal 実画面 (`modal-screen.tsx` / `form-sheet-screen.tsx`) はタブバーを
覆って全画面表示する modal/formSheet presentation の本来用途を維持する
ため、引き続きルート Stack 配下に残す。URL は (home) をグループルートに
することで全て維持され、`router.push("/navigation-patterns/...")` は
変更不要。

副次的な変更:
- ホーム画面の `SafeAreaView` を削除 (Stack ヘッダーが safe-area を管理)
- `(tabs)/_layout.tsx` 親 Stack 用に置いていた `<Stack.Screen.Title>ホーム</Stack.Screen.Title>`
  (d059338 で戻るボタンラベル復元用に追加) を削除。ナビゲーションパターン
  への push は (home) Stack 内に移ったため不要になった
- `NativeTabs.Trigger name="index"` を `name="(home)"` に変更
